Graphical representation.
The scene is done, complete with chosen materials for the two environmental objects and a spotlight (The scene is not done; it was determined that collision mesh to be created in and imported from Blender is an easier means of providing the player with level bounds than a programmed solution).

Player movement (And it's accompanying audio system) is to be done, next.