Marriage of character movement and an audio system ("Breathing" skeleton).
Player movement commands may be reinterpreted/nullified depending on the exhaustion of the first person character.

The system that monitors the character's physical effort is responsible for audial representation of the character's cycles of breathing.

Character movement is, in this regard, motivated by audio design. A passkey access-protection idiom has been implemented so that the Breathing System has special access to character movement mode control.