Breathing System.
A character may break into a sprint.
They must recover from the physical effort of a sprint.
Sprinting during recovery can be done once, though it exhausts the character.
If the character attempts to sprint while exhausted, the recovery duration is lengthened by the playback of an inhale.
The character will recover from exhaustion after a penalty recovery period.

The duration of any timed state is measured in the number of breath cycles. All states are represented, aurally.

Therefore, the first person character movement is aurally motivated.