Area Localisation of Drumming Effect.
The Area-localisation Sophisticated Sound Management system is comprehensively implemented.

Relative to distance and line-of-sight, variations of frequency and volume attenuation mimic atmospheric and surface absorption of sound.

The Drumming audio effect is localised to the trailer-top exterior - an AreaLocalisation actor pursues the first-person character within these bounds and achieves the effect of localising the point source to a broader surface-area.

The system's aural output may be isolated or toggled-off using the 'R' key.